[#46] Rename: modify file names to add Test suffix
